Adapt to wgpu 29 and egui 0.34 API changes
Covers the Cargo.lock/Cargo.toml dep bump: wgpu 29 reshaped
`InstanceDescriptor` and `PipelineLayoutDescriptor`, and egui 0.34
deprecated the `update`/panel-on-context flow in favor of a `ui`-based
API. Also drops a stale `gpu-alloc` profile entry and the obsolete
claude_auto_impl scripts.

+++ imaginarium/src/ops/contrast_brightness/cpu.rs
@@ -1576,7 +1576,7 @@ mod tests {
 
             // Check alpha bytes for each pixel
             for row in 0..4 {
-                let row_start = row * input.desc().stride as usize;
+                let row_start = row * input.desc().stride;
                 for x in 0..16 {
                     let pixel_start = row_start + x * pixel_size;
                     let alpha_start = pixel_start + alpha_offset;
